+++ src/mailmanclient/docs/NEWS.rst
@@ -9,7 +9,7 @@ NEWS for mailmanclient
   support filtering the response by a list domain.
 * URL encode values in URL which are url unsafe. (Closes #44)
 * Add support to mass unsubscribe memebrs from a Mailing List. (Closes #43)
-
+* Add support to set a user's preferred address. (See !99)
 
 3.2.2 (2019-02-09)
 ==================